requirements.txt
# Base
python-dotenv>=1.0.0
pydantic>=2.5.2

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# MQTT
paho-mqtt>=1.6.1

# API
fastapi>=0.104.1
uvicorn>=0.24.0

# Processing
pandas>=2.1.3
numpy>=1.26.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Logging
loguru>=0.7.2

# Development
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
